Add Platform Logging Lib
