+++ libs/common/spec/intercept-console.ts
@@ -1,20 +1,1 @@
-const originalConsole = console;
-
-declare let console: any;
-
-export function interceptConsole(): {
-  log: jest.Mock<any, any>;
-  warn: jest.Mock<any, any>;
-  error: jest.Mock<any, any>;
-} {
-  console = {
-    log: jest.fn(),
-    warn: jest.fn(),
-    error: jest.fn(),
-  };
-  return console;
-}
-
-export function restoreConsole() {
-  console = originalConsole;
-}
+export { interceptConsole, restoreConsole } from "@bitwarden/logging";
