+++ libs/common/src/platform/enums/log-level-type.enum.ts
@@ -1,8 +1,1 @@
-// FIXME: update to use a const object instead of a typescript enum
-// eslint-disable-next-line @bitwarden/platform/no-enums
-export enum LogLevelType {
-  Debug,
-  Info,
-  Warning,
-  Error,
-}
+export { LogLevel as LogLevelType } from "@bitwarden/logging";
