+++ libs/logging/README.md
@@ -1,0 +1,5 @@
+# logging
+
+Owned by: platform
+
+Logging primitives
