+++ libs/logging/eslint.config.mjs
@@ -1,0 +1,3 @@
+import baseConfig from "../../eslint.config.mjs";
+
+export default [...baseConfig];
