+++ libs/logging/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "@bitwarden/logging",
+  "version": "0.0.1",
+  "description": "Logging primitives",
+  "private": true,
+  "type": "commonjs",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "license": "GPL-3.0",
+  "author": "platform"
+}
