+++ libs/logging/src/log-level.ts
@@ -1,0 +1,8 @@
+// FIXME: update to use a const object instead of a typescript enum
+// eslint-disable-next-line @bitwarden/platform/no-enums
+export enum LogLevel {
+  Debug,
+  Info,
+  Warning,
+  Error,
+}
