+++ libs/logging/src/logging.spec.ts
@@ -1,0 +1,8 @@
+import * as lib from "./index";
+
+describe("logging", () => {
+  // This test will fail until something is exported from index.ts
+  it("should work", () => {
+    expect(lib).toBeDefined();
+  });
+});
